+++ src/ts/wysiwyg/processKeydown.ts
@@ -381,6 +381,16 @@ export const processKeydown = (vditor: IVditor, event: KeyboardEvent) => {
                 return true;
             }
         }
+
+        if (isCtrl(event) && event.shiftKey && event.key === ".") {
+            // 插入 blockquote
+            range.insertNode(document.createElement("wbr"));
+            blockquoteElement.innerHTML = `<blockquote data-block="0">${blockquoteElement.innerHTML}</blockquote>`;
+            setRangeByWbr(vditor.wysiwyg.element, range);
+            afterRenderEvent(vditor);
+            event.preventDefault();
+            return true;
+        }
     }
 
     // h1-h6
@@ -499,7 +509,7 @@ export const processKeydown = (vditor: IVditor, event: KeyboardEvent) => {
     if (taskItemElement) {
         if (isCtrl(event) && event.shiftKey && event.key === "j") {
             // ctrl + shift: toggle checked
-            const inputElement = taskItemElement.firstElementChild as HTMLInputElement
+            const inputElement = taskItemElement.firstElementChild as HTMLInputElement;
             if (inputElement.checked) {
                 inputElement.removeAttribute("checked");
             } else {